Experimental ncon (#652)
* fixing tests

* added a new ncon

* fix bug

* remove and add files

* remove prints add ncon_network

* add ncon_network back
